Development Enviornment & Project Structure
The internals for setting up a development enviornment, creating the neccessary folders and importing the neccessary repos.

Some work to be done on the initial store so that it renders something akin to a store on completion.

Framework files are symlinked the framework folder in your project root. You should make changes to the framework here.

Changes to the demo application should be made in the aero application itself.

We should all be aware of which concerns belong to which project.